I added the code for the quasi-normal step.
I added the code for the quasi-normal step.  Hopefully, this will serve
as the pattern for the rest of the augmented system solves.  In order to
accomplish this, I added a generic operator for the augmented system.  I
also added a generic block diagonal preconditioner built from the pieces
that the user provides.  It's this preconditioner that we'll build the
schur style preconditioner that Denis has been researching.  In
addition, I needed to add a couple more parameters to tune the behavior
of the augmented system solve.  I also added memory allocation and
restarting to a cached version of g(x) since we seem to compute it
often.  This matches the precomputation of h(x) for inequality
constraints.