I started adding code for the tangential...
I started adding code for the tangential subproblem.  More importantly,
I fixed a bug with the projected truncated, projected conjugate
direction code.  When we find our new iterate, it is important that we
update x with the projected direction and not the direction itself.
Otherwise, our iterate x will not lie in the desired projected space in
exact arithmetic.