I updated a comment on the Hessian...
I updated a comment on the Hessian modification for the interior point
method.  Basically, we require a symmetric modification of the Hessian.
This was in the code and labeled when the symmetric part is added.
However, the overall comments on the function did not include the extra
pieces.  This has been fixed.