There's some issue with libtool where the...
There's some issue with libtool where the dynamic library can't find the
components for blas.  Most likely, this occurs because we have a static
library for blas.  In any case, I do the unsafe thing and hard link
libpeopt.a when building the matlab libraries.  For the moment, this
appears to work.