This should be a working version of peopt in...
This should be a working version of peopt in Matlab.  It really needs to
be cleaned up and it's not as flexible as I'd like, but it appears to
work.  One irritating note, Matlab needs all libraries compiled with the
fPIC flag.  This causes us some problems with jsoncpp since that library
uses scons for its build system and we can't easily change the flags.
The easy way to get around this is to modify SConstruct and add the
-fPIC flag after something like Wall for the gcc-linux build.  If
there's a general way to handle this, I'd like to hear.