I now completely removed the init function...
I now completely removed the init function from the specification and do
not require the user to provide one.  Basically, we can get away with
this internally by doing a deep copy with mxDuplicateArray.  In
addition, I modified how we do memory management with the vector space
vectors.  Previously, we used the init function to implicitely allocate
memory.  In theory, MATLAB should free this memory for us, but I believe
it only does so when the program exits.  This is problematic for really
long runs or memory heavy algorithms such as quasi-Newton methods.  Now,
we've changed the vectors from mxArray* to a custom vector class that
contains an mxArray*.  By doing this, we're able to free this memory
using mxDestroyArray when the vector leaves scope.  In theory, this
means that we should be cleaning up our memory during the computation
and not just waiting until MATLAB exits.