I believe I found the source of the major...
I believe I found the source of the major outstanding memory leak.
Frequently, we call MATLAB functions from the mex file to routines the
user has called.  In these cases, the routines return some value we
need.  Now, typically, when we assign this value we do so to a variable
that has already allocated memory.  Since this assignment was done using
the MATLAB function mexCallMatlab, I assumed that this freed the
previously used memory.  This appears to not be the case.  Instead, we
have to manually free the memory from the output variables first and
then call mexCallMatlab in order to not leak.  From some simple tests,
it appears to have stopped the growth of memory used by MATLAB.