Previously, peopt could not safely call peopt...
Previously, peopt could not safely call peopt in one of its functions.
The reason for this is that the mxArray pointers that contained the
functions for things like the inner product were being cleared when the
inner peopt iteration exited.  This has been fixed.  Though, it's
possible that a small amount of memory (a few bytes) may be leaked.  In
addition, if the vector space is different in the inner iteration than
the outer, there may also be problems.