I added the ability to use a preconditioner...
I added the ability to use a preconditioner for truncated-CG.  For our
present use cases, we need the preconditioner to know the current
iterate in order to compute correctly.  However, this is not how we
typically call the preconditioner.  Typically, the preconditioner is
just something that takes in a vector and returns another vector.
Hence, it has no information about the current iterate.  In order to
communicate this information, we create a reference internally to the
preconditioner, which is held private.  This mechanism is difficult to
work with when hooking into Matlab since there's not a great way to
implicitely hook the current iterate to the preconditioner.  As such, we
have a hack to make it work where the Matlab interface can provide a
function with the current iterate.  Internally, inside of peopt, we then
do the right thing.  In the long term, it may be worthwhile to change
the preconditioner in peopt to take a base argument so that it can be
provided with the current iterate.