I made a few changes.
I made a few changes.  First, I updated the quasi-Newton methods so that
they use the updated memory management routines.  Second, I modified how
we check the size of the trust-region step for termination.  In the case
where we have a singular preconditioner for truncated-CG (generally a
bad idea, but it happens), it's possible that the norm of our step does
not reduce.  This occurs since we only check the preconditioned norm
when running truncated-CG.  I have sort of a hack that checks this right
now, so more thought needs to be given to it.  Finally, I rearraged the
steps in checking the predicted versus actual reduction.  Basically, we
want to do as much work at the current iterate as possible in order to
play more nicely with caching in the case of reduced-space formulations
and the solution operator.  By rearranging a few calls, we can reduce
everything to one solution operator call per iteration.  Before, we had
three.