I added the python wrappers for peopt.
I added the python wrappers for peopt.  Right now, they are incomplete,
but for the most part everything is there with some amount of error
checking.  The example that I'm putting in exercises the diagnostic
checks.  With a few more lines of code, I can turn on the actual
optimization.  At the moment, there appears to be a segfault when python
exits.  Somewhere, one of the python functions that I link to gets
dereferenced too many times.  I need to figure out where.  Finally, I
need to update the make system.  Notably, the shared library that I'm
making in this directory shares the same name as the main peopt library.
That's bad.  In theory, I could just change the name of the python bits,
but they should probably just be included in the regular peopt library.
That necessitates updating the build process, which will take some time.