I made a couple of changes to the code.
I made a couple of changes to the code.  First, I modified the object
based implementation of the operators by changing the order of the
arguments of the ternary operators.  Previously, we had y.prod(x,z)
where z is the return.  Now, we have z.prod(x,y) where z is the return.
This makes intuitively more sence to me.  Second, I eliminated a number
of temporary variables by realizing that the order of operations means
that the argument will be fully evaluated before the assignment.
Basically, we need to overwrite the pointer held in some object, but we
need this pointer in order to do the calcuation.  The calculation is
guaranteed to hold first, so we can calculate and then reassign the
variable in one go.  Next, I reworked the exception handling scheme.
Honestly, the c-api exception stuff was acting a little flaky.  It was
going off every other time I ran the program, which was weird.  Instead,
I modified the return from the c file so that it always returns a tuple.
If the second element in the tuple is None, then everything exited
correctly.  If it is a string, there's a problem and an exception should
be thrown.  Now, the exception is throw on the python side of the
interface code.