I've a working cmake build sans the python pieces.
I've a working cmake build sans the python pieces.  The only real hangup
right now is that the built-in BLAS and LAPACK find package utilities do
not work on my machine.  Basically, when using ATLAS, liblapack depends
in libcblas, which is not found by Kitware's find routines.  Rather than
using these, I put in a basic functionality that looks for libblas and
liblapack.  Most likely, this is not going to work, but then the user
can specify the appropriate libraries by hand.  I'm open to suggestions
if there's a better way to do this.  There's a huge number of
combinations for linking BLAS and LAPACK, so I don't think this issue is
unique to this library.  Other than that, the cmake menus look great and
we dynamically compile and install well.