This should be a working version with the...
This should be a working version with the regular peopt as well as the
python wrappers.  As a note, at the moment, we require shared versions
of the libraries for BLAS, LAPACK, and jsoncpp.  Basically, we produce
three libraries: libpeopt.a, libpeopt.so, and libpeopt_complete.so.  The
libraries libpeopt.a and libpeopt.so depend on json, BLAS, and LAPACK,
but they do not include these libraries.  The reason for this is that
it's perfectly plausible that the user would use these libraries on
their own.  If these libraries were rolled into ours, then there would
be conflicting symbols.  Nevertheless, this causes a problem when
linking to Python and potentially other languages such as MATLAB.  Here,
all of the libraries need to present.  There are a couple of ways to
accomplish this.  Either, all of the libraries need to be loaded in by
the user, or we need to roll a mega library that contains all of the
information.  For Python, this need and desire is a little more stark.
Basically, we need to tell python to load in peopt and we do so through
a Python command.  Since this library depends on BLAS, LAPACK, and
jsoncpp, these libraries need to be loaded as well.  Now, at the moment,
we take care of the commands to load in peopt and hide them inside our
module.  Doing this same trick, we can load in jsoncpp as long as we
assume that the library can be find on the LD_LIBRARY_PATH.  This is
much more difficult for BLAS and LAPACK since they don't have standard
names.  Now, we do find the names during our cmake configuration, but
we'd have to have some macro that extracted them into our Python module
if we want to be nice and load them on our own.  Otherwise, we'd have to
expect that our user would load the appropriate files and that seems
unreasonable.  As such, as a compromise, we create this mega library
called libpeopt_complete.so, which contains all dependencies.  As long
as we only use this library for hooking into MATLAB or Python, we should
be safe from the duplicate symbol problem.