I overhauled the MATLAB code so that we no...
I overhauled the MATLAB code so that we no longer require m4 macros in
order to create the vector spaces.  In theory, this should make
everything cleaner and easier to debug.  I also now return the solution
and the Lagrange multipliers.  Previously, we just returned the
solution.  Finally, I updated all of the examples to make sure of the
new interface.