I replaced the previous truncated-cg code with...
I replaced the previous truncated-cg code with a newer truncated
conjugate direction code.  Part of the motivation for this is to prepare
for the equality constrained algorithm, where this will be required.  In
addition, this allows us to abstract out the Krylov solver into the
linear algebra code so that it could be tested individually.  Finally,
for some problems, there can be an issue with stability.  Namely, it
turns out that truncated conjugate directions requires fewer Krylov
solvers than truncated conjugate gradient.  Now, by default, the number
of stored vectors is set to one, which is in some sense the cg
compatibility mode.  However, this can be modified and turned up.  As
far as the penalty for doing this, I'd have to sit down and count, but
there may be 1 or 2 more additional vectors required for this algorithm
to work.  As such, there's a slightly higher memory cost, but it's small
relative to the overall algorithm.  Finally, I modified the semantics of
this function so that we look at the norm of the step directly when
determining whether or not we're outside the trust-region radius.
Previously, the trust region we used was in the preconditioner norm || .
||_{M^{-1}}.  This the commonly stated algorithm from the big blue Conn,
Gould, and Toint book.  Now, in my opinion, the reason they did this is
that they frequently use this preconditioner in order to handle scaling
differences between variables.  It turns out, this is also a good thing
to do when preconditioner CG.  However, in our code, we work with a
general inner product, which means that it's more natural to take care
of scaling problems in the inner-product rather than the preconditioner.
This decouples the preconditioner in such a way that we can focus on it
being a good precondioner for the linear system.  In addition, this
allows us to use non positive definite precondioners.  Since we already
have a stopping criteria for negative curvature, we should be able to
use, essentially, whatever.  I've not tested it in the most general
case, but there are instances in parameter estimation where it's
worthwhile to precondion with a positive-semidefinite preconditioner.
Basically, if we use an optimization algorithm to solve a nonlinear
system of equations and we have the Jacobians lying around and we use
Gauss-Newton, the method flies, but the Jacobians can and frequently
will be degenerate.  This occured when solving nonlinear elasticity.
Anyway, the old code had a hack in order to handle this.  With the new
setup, it works just fine in a more natural way.