This is a major rework of the interior point...
This is a major rework of the interior point method.  Basically, the old
method had a number of problems, which may have also included
correctness.  The correct way to handle the interior point paramter is
to estimate it at every iteration and then the Newton system aims for
sigma times mu where sigma is some sort of centering paramter.  When
sigma=1, we do a pure centering step.  When sigma is less than 1, we
move toward the solution.  Now, I've put in three different interior
point methods.  The default is to do an agressive primal dual interior
point method where the primal and dual steps are deocupled.  This is
typically what is done in linear cone codes.  Next, I put in a similar
code to what we had before where we link the primal and dual steps.
This is somewhat more conservative, but it appears to work well.
Finally, I put in a primal log-barrier method.  Honestly, this this is
behaving terribly.  Perhaps there's some sort of error, but I find it
incredibly difficult to get an accurate estimate of the optimality
conditions.  Essentially, a primal barrier method is like a primal dual
method, except that we solve a closed form solution for the lagrange
multiplier at every step.  I don't find that this solution method is
very robust or accurate, which leads the method to find a pretty good
solution, but indicate that the gradient has not converged.  Perhaps it
requires some tuning; I'm not sure.  In addition, I cleaned up some of
the code and added some new examples that exercise the interior point
method for linear, second-order cone, and semidefinite programs.
Finally, I started changing over the restart routines to use the method
I developed in the equality constrained branch.  It's necessary now
since the interior point code allocates more than just vectors in the Z
space.  It now allocates vectors in the X space as well.