I made some non-backward compatible changes to...
I made some non-backward compatible changes to the code.  Namely, in the
past, we used the name s for the trial step.  Now, we use dx.  The
reason for this change is that we need to start modifying our notation
for the introduction of equality constraints.  This also led to the
elimination of rho, the actual vs. predicted reduction, and its
replacement by two values ared and pred.  In addition, I fixed a bug
associated with the Hestenes-Stiefel version of nonlinear conjugate
gradient.  The formulas for HS dictate that the momentum paramter is
dependent on the previous step.  Now, since we have both trust-region
and line-search methods, we always scale our steps by alpha.  This
allows us to update the step with x+dx instead of having different codes
between trust-region and line-search that would use x+dx and x+alpha dx
respectively.  The consequence for this is that we actually need to
rescale the previous search direction in the HS formula.  In addition, I
was running into some problems with nonlinear-CG where we didn't have a
valid search direction.  Since we can not do an exact line-search and we
never check the strong-Wolfe conditions, it is possible that we lack a
descent direction using the available formulas.  In order to combat
this, we now do a hard check for descent.  If we do not have a descent
direction, we simply flip the direction so that we do.  Finally, I added
some new unit tests to make sure that nonlinear-CG performs in the
expected way on quadratic problems.