I modified the code so that we no longer...
I modified the code so that we no longer depend on unsigned ints, but
rather depend on a typedef Natural, which is defined to be
std::vector<double>::size_type inside of linalg.h.  Now, in terms of
replaces integers, we have a couple of options.  One, which I originally
modified the code to, is to template all our functions on the Natural
type.  This includes our vector spaces.  On the outset, this seemed like
a reasonable thing to do, but there were some quirks.  Namely, it seemed
odd to have to template the vector space on a Natural type.
Mathematically, a vector space consists purely as a relationship between
a Field and a Set.  Therefore, enriching this with some sort of Natural
number seemed strange.  In addition, in a practical sense, why do we
care if we have a generic Natural?  Mostly, it's because we want to be
able to index correctly regardless if we're on a 32-bit or a 64-bit
machine.  In the STL, this question has already been handled.
Basically, if we index a vector, we must do so using the
std::vector<T>::size_type type.  This is automatically adjusted between
32-bit and 64-bit platforms and is guaranteed to be an unsigned integer.
Therefore, it seemed to make more sense to rely on this than something
else.  Now, there are some problems with these scheme, but I don't think
they're unique to templating or just doing the typedef.  First, file IO
is still a major pain and there doesn't seem to be a perfectly generic
way to do it.  For example, when we do the json i/o, the question is
whether or not we should read in 32 bit ints or 64 bit ints.  Frankly, I
don't know, so what we do is just tell it to read in 64 bit ints and
then build the integer type from the result.  If we're working on a 32
bit machine and we read something too big, there may be an overflow.
We'll have to live with that for now.  In addition, the diagnostics
aren't completely generic at the moment.  Basically, we need a way to
convert our real type as well as our natural type to something that can
be printed.  At some point, this code needs to be fixe since it's
specialized right now to doubles.  Finally, the interaction with BLAS is
note templated on generic integer types.  It's not clear to me if we
could actually do this since I think the basic BLAS routines are hooked
to 32-bit ints.  Certainly, we could write out own BLAS routines to fall
back on in the case where we don't have the correct library.  In any
case, the structure of the code is such that we could template on the
Natural type if necessary and do so quickly.  I don't really think
that's the best option, though, since it would affect our ability to
easily interact with the STL.