I did a large amount of cleanup on how...
I did a large amount of cleanup on how integers and natural numbers are
handled, the default vector spaces, and the interaction with BLAS.
Previously, I had set the Natural type to be a typedef of vector
<double>::size_type.  Now, I've altered that and I set Natural to size_t
and Integer to ptrdiff_t out of cstddef.  The reason for this is to get
a standard definition of either unsigned or signed integers that should
adapt corectly between 32-bit and 64-bit machines as well as windows.
Part of the other motivation for doing this was to clean up the
interface between this code and BLAS and LAPACK.  The native BLAS
routines require us to pas in the INTEGER type.  My understanding is
that the INTEGER type in Fortran 77 is not fixed to be a certain number
of bytes, but instead depends on the architecture of the computer.
Hence, it may be 4 bytes or it could be 8 bytes (or something
different).  My guess, and hope, is that the size of this integer should
be the same as ptrdiff_t.  For most common architectures and compilers,
I believe this to be the case.  Anyway, I recoded all of these routines
to be parameterized on the new Integer type in the hope that our code is
more compliant on 64-bit systems.  Next, in a very similar manner to
wrapping all constants in Real(), I've tried to wrap all our integers
with the appropriate Natural or Integer constructor.  In theory, C++
should do all of the automatic typecasting for us, but I frankly don't
like automatic typecasting and would rather do so by hand in order to
better insure that we don't have type errors.  There is a material
difference between Integer and Natural and I prefer that difference to
be very visual.  This required some recoding of the vector spaces, which
became a little verbose.  In order to combat that, I introduced some new
functions to grab certain pieces of the vector without having to do
indexing.  For example, with a second-order cone program, we typically
have a vector that's of the form x=[x0;xbar] where xbar is a vector and
x0 is a scalar.  For many operations, we need either x0 or xbar.  In the
previous style, we'd index this as x(blk,1) or x(blk,2).  Now, we index
things as x.naught(blk) and x.bar(blk).  This prevents us from having to
wrap the indices with the appropriate type and makes the code a bit
cleaner.  Note, in the course of all of these conversions and
interfacing with BLAS, we frequently need to convert a Natural to an
Integer.  In theory, for really be vectors, this could cause an
overflow, which will lead to errors that we do not catch.  In reality,
the chance of that happening is pretty small, but beware that weird
behavior could occur on 32-bit systems with really big problems.  I'm
not sure if we should catch that right now or not.  Perhaps it's a
safety thing that we could add in the future.