The json code has some trouble with...
The json code has some trouble with determining the correct type between
our natural type and its internal types.  We now force type cast these
numbers to UInt64.  In the json code, if we're on a 32 bit machine, this
falls back to 32 bit ints.  Now, the code where we form the json tree
matches the code where we read the json file.