I introduced a new Krylov solver for solving...
I introduced a new Krylov solver for solving the trust-region subproblem
or the Newton Krylov method.  Basically, it's a variant of MINRES, which
should, in theory work well.  In practice, I'm still trying to figure it
out since it's not clear that it always gives a descent direction.  At
this point, I wouldn't recommend using it until this is all sorted out.
In addition, I discovered that the total number of Krylov iterations is
not being reported properly.  This was likely a bug that was introduced
when the conjugate direction code was added.  In any case, this has been
fixed.