I fixed the truncated GMRES routine and made...
I fixed the truncated GMRES routine and made it a true truncated MINRES
routine with the same calling signature and truncated conjugate
direction.  The real difference between the two lies in how the
preconditioner is applied as well as in the memory required and how we
find our iterates.  Previously, we cheated a bit and did a left
preconditioning similar to GMRES.  This is not really what we need.
Rather, we need to restrict ourselves to positive semidefinite
operators, which requires us to use a different formulation.  In
addition, we previously still needed to solve a triangular (band)
system.  This has been eliminated a new scheme has been added to update
the iterate at each iteration.  This actually makes it a bit easier to
implement the truncated pieces since the new step is readily available
at each iteration.  Finally, I modified the routine to only hold exactly
the required number of vectors.  This does mirror the conjugate
direction code in that we have the ability to over orthogonalize, but it
does not by default.  In any case, it appears to be working and I added
a number of tests to the unit test to verify.  Given that it now uses
the same calling sequence as conjugate direction, it'd probably be best
if we figured out a way to unify their unit testing.