I rearranged the computations inside the code...
I rearranged the computations inside the code that checks the
trust-region step.  Basically, in a reduced space method that uses
caching in the solution operator, this allows us to delay evaluating a
function at a new point.  This means that the caching works more
effectively and we should only have to evaluate the state once per
iteration.  Previously, because things were in a less than optimal
order, we had 3 state solves since we evaluated at the current point, a
new point, and then back at the current point.  Note, this only affects
codes that cache automatically.  If we were caching by hand, this
probably wouldn't have been an issue.