I modified truncated-MINRES so that it now...
I modified truncated-MINRES so that it now traps NaNs in both the
preconditioner and the operator.  There are a couple of reasons why this
occurs.  First, if the preconditioner is not positive definite, then
it's possible that the normalizing factor in Gram-Schmidt is zero, which
is bad.  In addition, it could be that either the operator or the
preconditioner is incorrectly computed or comes for a piece of the code
that generates a NaN.  For example, in a reduced space method, we may
violate the CFL condition between iterations and then generate a NaN.
Hence, it's important to catch it.  Now, conjugate direction already
does this, but I'm not entirely convinced it does so early enough.  So
far, our codes are working, but it appears a little bit odd.  We should
investigate this further and add some tests that creates this behavior.