I modified both truncated conjugate direction...
I modified both truncated conjugate direction as well as conjugate
MINRES in order to better handle instabilities.  Basically, it's
possible that either the preconditioner or the Hessian will return a NaN
due to a variety of reasons.  Most notably, this occurs when using an
explicite time integrator in a reduced space method for which the time
step violates the CFL.  In these cases, we mostly handled things
properly in the past, but I think the new code does a better job at
that.  In order to accomplish this, we try to detect the instability
earlier in the Krylov method and then skip a number of steps that don't
need to be computed if we exit.  In addition, we are more careful now to
not use potentially invalid steps, which are normally based on the
Krylov vectors, when figuring out the truncated step.  Rather, the
truncation step uses the last valid step we have calculated in order to
do the truncation calculation (this normally means scaling the iterate
all of the way to the trust-region).  Anyway, I need some unit tests
still to test this new code, but our old tests still pass and some other
problems appear to react well to the change.