I fixed some of the diagnostic output to...
I fixed some of the diagnostic output to mirror the variables actually
in the code.  Recall, in preparation for both inequality and equality
constraints, I had to change our new step from s to dx.  In doing so, I
forgot to update the header for the diagnostic output, which is what
this check-in is doing.