I modified a variable initialization to be NaN.
I modified a variable initialization to be NaN.  As a general rule, I've
tended to not initialize variables that are explicitely set later
because if I make a mistake in their use, debug mode tends to pick them
up quickly as these numbers are set to bad values.  Technically, it
would be safer to initialize them to NaN as that would force a bad
situation, which is easier to debug, which is what I do here.  It is not
good to initialize them to a safe value as it makes it harder to debug.