I modified the code so that we physically...
I modified the code so that we physically prevent someone from trying to
allocate what really amount to templated namespaces.  There is no reason
to instantiate these classes ever, so this change should be safe.  In
addition, I opened up some of our packaged enumerated types and made
them namespaces as opposed to classes.  Basically, we need a
construction so that our enumerated types don't conflict.  Sometimes, we
want to use the same names.  In addition, we need a couple of associated
functions with most enumerated types such as converting them to a
string.  Previously, this was all packaged inside of a class.  There's
nothing wrong with doing this, but if we can package it in a namespace,
that's probably better.