I fixed the routine that projects the gradient...
I fixed the routine that projects the gradient of the Lagragian plus the
Hessian applied to dx_n.  This is the linear piece of the quadratic
model that's sent to truncated conjugate direction.  Basically, there's
a notational difference between what's used internally in this code and
with the tech report that Denis, Miguel, and Matthias wrote on the
inexact composite step algorithm.  In their tech report, g = grad L(x,y)
+ H dx_n = grad f(x) + g'(x)* y + H dx_n.  In our code, we use g = grad
L(x,y).  When translating the algorithm, it's important to note this
difference.  In addition, I used to have code to compute this piece as
an operator.  I removed that and made it a regular routine.  Unlike most
of out other routines, the result of this computation W (grad L(x,y) + H
dx_n) is *not* currently cached in the state, but rather returned by
reference.  Really, we just need this quantity once, so it's not clear
to me that it's worth the effort to carry it around.