I modified our two primary Krylov methods to...
I modified our two primary Krylov methods to take in a new parameter,
the center of the trust-region.  Essentially, this allows us to change
the center of where we calculate the trust-region constraint.  This is
important for the equality-constrained algorithm.  In the unconstrained
algorithm, we simply set the center to be 0.  In addition, I added two
new unit tests.  These are a little tricky to generate in an independent
way since, as far as I know, we need to know exactly how each algorithm
reacts in order to get it to hit the shifted trust-region in a known
location.  Mostly, we take advantage that CG will move in the steepest
descent direction in the first iteration, which is just the RHS.  In
MINRES, we know that it will choose an optimal point in the subspace
generated by the first Krylov vector, which is also just the RHS.