I've been futzing with some of the notation.
I've been futzing with some of the notation.  Basically, we have two
different tangential steps: the one generated by our truncated Krylov
method and the one generated after projecting this step.  Mostly, we
need a way to keep track of which is what.  In the past, I labeled these
dx_t for the one generated by the Krylov method and then dx_tnull for
the one after projection.  This confused me, so I changed it.  Now, the
one from the Krylov method is called dx_t_uncorrected and the one after
the projection is dx_t.  We'll try that for awhile until I get confused
again.