I added caching for some additional quantities.
I added caching for some additional quantities.  We now store the value
of g'(x)dx_n+g(x).  This is used in the predicted reduction and the
residual predicted reduction.  Since we may need these quantities
several times due to the uncorrected dx_t changing, it makes sense to
cache them.  In addition, I added caching for both || g(x) || as well as
|| g'(x)dx_n + g(x) ||.  These quantities are needed for the computation
of the penalty paramter as well as in the predicted reduction and
residual predicted reduction.  As with g'(x)dx_n + g(x), these items are
repeatedly required due to the uncorrected dx_t changing, so we need to
cache them.