I added in most of the remaining functions for...
I added in most of the remaining functions for the equality constrained
algorithm.  Right now, I'm trying to decide on the best way to code the
getMin function.  Basically, it's not all that different that the one in
the unconstrained code, so I'm trying to figure out if we should just
copy it or if there's a better way to reuse the code.  In addition, I
eliminated f_merit and model from the bundle of functions.  At the end
of the day, there's not too many reasons to keep around f, f_merit, and
model.  We don't really need the plain objective function ever.  All we
need is the merit function, which means that we can replace the
evaluation of f with the merit function.  In addition, we don't really
need the gradient or the Hessian-vector product, we need the gradient of
Lagrangian or the Hessian of the Lagrangian.  Therefore, it seems safe
to replace these bits and just encapsulate everything in f.  Now, this
means that there's a lack of consistency between f and its gradient.
Meaning, that a finite difference check will fail if we try to do one on
the modified functions.  However, we don't ever do that since our
diagnostics are run on the raw function and not the modified one.  As a
note, the inequality stuff does do some magic with the gradient since it
turns out for quasi-Newton methods, we do want either the raw gradient
or the gradient of the Lagrangian in order to build the Hessian
approximations.  In addition, I fixed a bug in the conjugate direction
code where x-x_cntr was not being updated properly, so we were not
terminating when we needed to, which led to infinite loops.