This is a compiling version of the equality...
This is a compiling version of the equality constrained algorithm.  It
hasn't been debugged yet, so there's probably some errors.  At this
point, we need to debug these errors, which is why I've added a new
example that contains a very simple equality constraint.  In addition,
we still need to write the json reader for the equality constrained
problem.