I hooked in the composite step function, which...
I hooked in the composite step function, which I was not doing before.
I addition, I added the equality constrained parameters to the JSON
reader.  In doing this, I realized that having a separate variable for
xi_all didn't makes sense since it wasn't clear how to update the other
values (xi_qn, xi_pg, etc.) when this value is set.  As such, I changed
it to be a function that sets everything.  Therefore, we can start out
by setting xi_all and then change the individual parameters.  For the
output to JSON, xi_all we be by default xi_qn.