I fixed a number of bugs, mostly relate to the...
I fixed a number of bugs, mostly relate to the GMRES manipulators, which
sets the linear system solve tolerance.  At this point, we can solve the
very simple equality constrained problem, but we require most extensive
tests in order to assess the behavior for more nonlinear problems as
well as inexactness.