I added a check for positive predicted...
I added a check for positive predicted correction to both the code for
calculating the penalty parameter as well as the code that refines the
tangential step.  Basically, due to either a bad implementation or some
odd numerical difficulties, it may be that our tangential step is bad
and that we calculate a negative predicted reduction.  In theory, this
should never happen, but numerics are they way they are.  Anyway, in
this situation, the check for calculating the penalty parameter doesn't
make sense and no amount of refining the tangential step, meaning making
sure that the tangential step is in the nullspace of the constraints,
can fix this problem.  In theory, the tangential step could be so far
off the nullspace that a better projection could fix this, but I'm not
sure how to deal with this case yet.  Anyway, the hammer approach to
this is to just bail on our attempt and shrink the trust-region.
Eventually, the trust-region should be small enough where we get
positive predicted correction.  Since our current code to check the
actual vs. predicted correction does this well, we allow this part of
the code to make the necessary adjustments.