I actually added the stopping condition for...
I actually added the stopping condition for making sure that the norm of
the constraint is small.  Evidently, our examples haven't been hitting
this problem yet, so we need to test this more carefully.  I also moved
where we calculate the update cached version of g(x) and its norm to the
CompositeStepManipulator, since it makes sense to update this value as
soon as we update the step.  Finally, for the feasibility stopping
condition, this is a relative tolerance controlled by eps_constr.  This
can and will be problematic since many users try to use an initial
feasible solution.  If this is the case, then eps_constr needs to be
somewhere around 1 instead of its default value which is 1e-6.  Really,
I don't know of a good way around this since the scaling for the
constraint could be sort of funny and it's not clear what near zero
means without some sort of relative difference.